+++ examples/data-sources/foundry_model_version/data-source.tf
@@ -1,0 +1,8 @@
+data "foundry_model_version" "classifier" {
+  name    = "support-classifier"
+  version = "1"
+}
+
+output "classifier_id" {
+  value = data.foundry_model_version.classifier.id
+}