+++ examples/resources/foundry_model_version/import.sh
@@ -1,0 +1,2 @@
+# Model versions are imported as name/version.
+terraform import foundry_model_version.classifier support-classifier/1